feat(bpf-rs): kernel descriptor
Given an fd that represents a kernel data structure, we can query the
kernel (through the bpf_prog_query subcommand part of the bpf() syscall
interface), for which bpf programs are attached to this data structure.

This is useful in particular for trying to understand which bpf programs
are attached to a cgroup.

This is a MAJOR patch change

Signed-off-by: Milan <milan@mdaverde.com>